Add user avatar and menu to Index.cshtml
Added `avatarSrc` variable to `Index.cshtml` code to display the avatar of the authenticated user. Instead of the "Login" button, a button with an avatar is now used, which opens the user menu. The menu contains links to your profile and a “Login” button, which improves the interface and ease of use.