Removed an error when deleting an account
Improved the handling of the server response after deleting an account. Added a check for the presence of the `success` property in the returned data. Now, if the response is not successful or `data.success` is `false`, an error is thrown with the message from the server or with the default message. This allows for more accurate error handling and more informative messages to the user.