Added migration class for roles and users
Added a new migration class `AddManagerRole`, which defines a database model for the entities `SkillModel`, `User`, `UserInfoModel`, `UserSkillModel` and user roles. Included are properties and constraints for each entity, and relationships between them are defined. The migration also includes settings for identifying users and roles using `IdentityRole` and `IdentityUser`, with appropriate indexes and constraints.