Add conditions for displaying links in Index.cshtml
Changes have been made to `Index.cshtml` to display links depending on the user's authentication status and roles. A link to the "Employees" page has been added for authenticated users and a link to "Users" for users with the "Admin" role.
The link to "Users" has been removed from the general list for unauthenticated users.