Updated navigation and role checks in Index.cshtml
Changes have been made to the display of navigation elements.
Added checks for user authentication and their role. Now, during authentication, links to "Employees" and "Newbies" are displayed.
Added a check for the "Admin" role to display a link to "Users". Changed the logic for checking roles for access to actions for creating and editing skills.