Improved user management in UserInfoController
Added dependencies for user management and optimized the logic for getting the current user.
Updated methods for creating and editing users taking into account access rights.
Simplified processing of reviews and skills, added checks for the presence of rights when changing the main image.
Removed restrictions on roles for some methods for more flexible access control.