Added CreatedByUser fields to UserInfoModel
Changes were made to the migration file `20250306125813_AddCreatedByUserFieldInUserInfo.Designer.cs`.
Added new fields `CreatedByUserId` and `CreatedByUserName` to the `UserInfoModel` model to track the creator of the record.
Updated the `SkillModel`, `User`, `UserSkillModel` models and added the corresponding relationships between them, which improves the representation of users and their skills in the database.