Added ReviewModel class with basic properties
A new class ReviewModel has been added to the ReviewModel.cs file.
The class includes the properties Id, Text, Rating, CreatedAt, and two virtual properties UserInfo and User.
The constructor initializes CreatedAt with the current date and time, and the rest of the properties have a default value set to null!.