Updated display of user cards
Changed the way users are displayed in the list.
Removed old code blocks and added new ones with improved structure and styles.
The height of the user card has been changed from `min-h-[300px]` to `min-h-[330px]`.
Added a block for displaying the average user rating with a corresponding message in the absence of reviews.
Simplified the logic for displaying the "More details" button for authorized users.